MacGui: fix a couple nasty rounding errors.
The results could be off by almost 4% with 30fps sources. 


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@7154 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBRange.m
@@ -88,7 +88,7 @@ NSString *HBRangeChangedNotification = @"HBRangeChangedNotification";
     else if (self.type == HBRangeTypeFrames)
     {
         hb_title_t *title = self.title.hb_title;
-        int duration = (self.frameStop - self.frameStart) / (title->vrate.num / title->vrate.den);
+        int duration = (self.frameStop - self.frameStart) / (title->vrate.num / (double)title->vrate.den);
         return [NSString stringWithFormat: @"%02d:%02d:%02d", duration / 3600, ( duration / 60 ) % 60, duration % 60];
     }
